% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Downloads and processes the data from the API}
\usage{
getdata(origin, destination, year, classification)
}
\arguments{
\item{origin}{Country code of origin (e.g. "chl" for Chile)}

\item{destination}{Country code of destination (e.g. "chn" for China)}

\item{year}{The OEC's API ranges from 1962 to 2014}

\item{classification}{Trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing. The default is set to "6".}
}
\description{
Downloads and processes the data from the API
}
\examples{
# Run countries_list() to display the full list of countries
# Chile is "chl" and China is "chn"

# Download trade data from OEC's API (HS92 6 characters product lists)
# getdata("chl", "chn", 2014)
# is the same as
# getdata("chl", "chn", 2014, 6)

# Download trade data from OEC's API (HS92 8 characters product lists)
# getdata("chl", "chn", 2014, 8)

# Download trade data from OEC's API (SITC rev.2 4 characters product lists)
# getdata("chl", "chn", 2014, 4)
}
\keyword{functions}

