% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Creates a network for a given year}
\usage{
network(origin, destination, classification, year)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{year}{is the year and the OEC's API ranges from 1962 to 2014}
}
\value{
Creates an \code{HTML} file with a network visualization for a given year.
}
\description{
Creates a network for a given year
}
\examples{
network("chl", "chn", 6, 2014)
}
\keyword{functions}

