% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{network_comparison}
\alias{network_comparison}
\title{Creates a network of exports to see if new exported products have acquired a comparative advantage within a period of year}
\usage{
network_comparison(origin, destination, initial_year, final_year,
  classification)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{initial_year}{is the initial year and the OEC's API ranges from 1962 to 2014}

\item{final_year}{is the final year and the OEC's API ranges from 1962 to 2014}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}
}
\value{
Creates an \code{HTML} file with a network visualization that compares two given years to see if more exported products have acquired a Revealed Comparative Advantage (RCA > 1) within the period.
}
\description{
Creates a network of exports to see if new exported products have acquired a comparative advantage within a period of year
}
\examples{
# Visualize trade data from OEC's API (HS92 6 characters product list)
# for exports from Chile to China in the year 2014
# network_comparison("chl", "chn", 2010, 2014, 6)
# is the same as
# network_comparison("chl", "chn", 2010, 2014)
}
\keyword{functions}

