\name{sitc_rev2_2char}
\alias{sitc_rev2_2char}
\title{SITC (rev. 2) groups}
\docType{data}
\description{
SITC (rev. 2) groups. This file is used to create the visualizations.
}
\usage{hs92_2char}
\format{
  A data frame with 36 observations on the following 2 variables.
  \describe{
    \item{\code{group}}{Contains the SITC (rev. 2) groups (e.g. machinery, electronics, etc)}
    \item{\code{group_id}}{Contains the associated codes of every group (e.g. machinery is 10)}
  }
}
\examples{
sitc_rev2_2char
}
\keyword{datasets}
