% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_interval.R
\name{treemap_interval}
\alias{treemap_interval}
\title{Creates a treemap for a given given period of years}
\usage{
treemap_interval(origin, destination, variable, classification, first_year,
  last_year, interval, depth)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{variable}{is the variable to visualize and it can be "imports", "exports" or "exchange" (trade exchange)}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{first_year}{is the first year and the OEC's API ranges from 1962 to 2014}

\item{last_year}{is the last year and the OEC's API ranges from 1962 to 2014}

\item{interval}{is an optional parameter to define the distance between years (by default set to 1)}

\item{depth}{is an optional parameter that can take values "0" (group's detail) or "1" (product's detail), by defaults its set to 1}
}
\value{
Creates an \code{HTML} file with a treemap visualization for a given period of years.
}
\description{
Creates a treemap for a given given period of years
}
\examples{
# treemap_interval("chl", "chn", "exports", 6, 2011, 2014, 2)
}
\keyword{functions}

