% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_interval.R
\name{getdata_interval}
\alias{getdata_interval}
\title{Downloads and processes the data from the API}
\usage{
getdata_interval(origin, destination, classification, first_year, last_year,
  interval)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{first_year}{is the first year and the OEC's API ranges from 1962 to 2014}

\item{last_year}{is the last year and the OEC's API ranges from 1962 to 2014}

\item{interval}{is an optional parameter to define the distance between years (by default set to 1)}
}
\description{
Downloads and processes the data from the API
}
\examples{
# Run countries_list() to display the full list of countries
# Chile is "chl" and China is "chn"

# Download trade data from OEC's API (HS92 6 characters product lists)
# for the years 2010, 2012 and 2014
# getdata_interval("chl", "chn", 6, 2011, 2014, 2)

# Download trade data from OEC's API (HS92 8 characters product lists)
# for the years 2010, 2012 and 2014
# getdata_interval("chl", "chn", 8, 2011, 2014, 2)

# Download trade data from OEC's API (SITC rev.2 4 characters product lists)
# for the years 2010, 2012 and 2014
# getdata_interval("chl", "chn", 4, 2011, 2014, 2)
}
\keyword{functions}

